+++ client/src/hooks/api/useContact.js
@@ -1,13 +1,16 @@
 import { useMutation } from "react-query";
 
 async function postContact(payload) {
-    const response = await fetch(`${process.env.API_URL}/API/contact`, {
-        method: "POST",
-        headers: {
-            "Content-Type": "application/vnd.api+json",
-        },
-        body: JSON.stringify(payload),
-    });
+    const response = await fetch(
+        `${process.env.APP_CLIENT_API_URL}/API/contact`,
+        {
+            method: "POST",
+            headers: {
+                "Content-Type": "application/vnd.api+json",
+            },
+            body: JSON.stringify(payload),
+        }
+    );
 
     const data = await response.json();