+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 Changes are tracked as incremental revisions.
 
+## [r9] - 2020-07-12
+
+-   Add mail transport support via [nodemailer](https://nodemailer.com/about/)
+-   Improve client .env support with port option
+
 ## [r8] - 2020-07-12
 
 -   Add server app tests